+++ crates/ruma-events/src/enums.rs
@@ -17,6 +17,12 @@ event_enum! {
         "m.push_rules" => super::push_rules,
         "m.secret_storage.default_key" => super::secret_storage::default_key,
         "m.secret_storage.key.*" => super::secret_storage::key,
+        #[cfg(feature = "unstable-msc2545")]
+        #[ruma_enum(ident = AccountImagePack, alias = "m.image_pack")]
+        "im.ponies.user_emotes" => super::image_pack,
+        #[cfg(feature = "unstable-msc2545")]
+        #[ruma_enum(ident = ImagePackRooms, alias = "m.image_pack.rooms")]
+        "im.ponies.emote_rooms" => super::image_pack,
     }
 
     /// Any room account data event.
@@ -130,6 +136,9 @@ event_enum! {
         "m.room.topic" => super::room::topic,
         "m.space.child" => super::space::child,
         "m.space.parent" => super::space::parent,
+        #[cfg(feature = "unstable-msc2545")]
+        #[ruma_enum(ident = RoomImagePack, alias = "m.image_pack")]
+        "im.ponies.room_emotes" => super::image_pack,
         #[cfg(feature = "unstable-msc3489")]
         #[ruma_enum(alias = "m.beacon_info")]
         "org.matrix.msc3672.beacon_info" => super::beacon_info,